♻️ move Lead page component to shared folder (#601)
