+++ web/pages/news/[slug].tsx
@@ -15,7 +15,7 @@ import NewsBlockContent from '../../common/NewsBlockContent'
 import { menuQuery } from '../../lib/queries/menu'
 import { footerQuery } from '../../lib/queries/footer'
 import HeroImage from '../../pageComponents/shared/HeroImage'
-import Lead from '../../pageComponents/news/Lead'
+import Lead from '../../pageComponents/shared/Lead'
 import RelatedContent from '../../pageComponents/news/RelatedContent'
 import LatestNews from '../../pageComponents/news/LatestNews'
 import { Icon } from '@equinor/eds-core-react'
